+++ module-infra/src/main/java/com/checkping/infra/repository/project/taskboard/TaskBoardRepository.java
@@ -1,4 +1,4 @@
-package com.checkping.infra.repository.project;
+package com.checkping.infra.repository.project.taskboard;
 
 import com.checkping.domain.board.TaskBoard;
 import com.checkping.domain.board.TaskBoard.BoardCategory;
