+++ module-infra/src/main/java/com/checkping/infra/repository/project/taskboard/TaskBoardStore.java
@@ -1,4 +1,4 @@
-package com.checkping.infra.repository.project;
+package com.checkping.infra.repository.project.taskboard;
 
 import com.checkping.domain.board.TaskBoard;
