+++ module-service/src/main/java/com/checkping/service/project/TaskBoardServiceImpl.java
@@ -5,8 +5,8 @@ import com.checkping.dto.TaskBoardRequest;
 import com.checkping.dto.TaskBoardRequest.SearchCondition;
 import com.checkping.dto.TaskBoardResponse;
 import com.checkping.dto.TaskBoardResponse.TaskBoardDto;
-import com.checkping.infra.repository.project.TaskBoardReader;
-import com.checkping.infra.repository.project.TaskBoardStore;
+import com.checkping.infra.repository.project.taskboard.TaskBoardReader;
+import com.checkping.infra.repository.project.taskboard.TaskBoardStore;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
